CSSTUDIO-1936: refactor HTML preview component
updated component api to use commonmark source directly, instead of callbacks
removed duplicate Remarkable configs, replaced with preview component
updated EntryEditor exports to re-export instead of putting all into index; improved readbility
